requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
openai
anthropic
requests
gunicorn
